Rename `convert` to `process`
More consistent with its "main" function's name
