+++ src/gfx/main.cpp
@@ -24,7 +24,7 @@
 #include "platform.h"
 #include "version.h"
 
-#include "gfx/convert.hpp"
+#include "gfx/process.hpp"
 
 using namespace std::literals::string_view_literals;
